# Conflicts:
#	backend/Draw/src/main/java/com/yehah/draw/domain/child_work/controller/ChildWorkController.java
#	backend/Draw/src/main/java/com/yehah/draw/domain/child_work/entity/ChildWork.java
#	backend/Draw/src/main/java/com/yehah/draw/domain/child_work/service/ChildWorkService.java
#	backend/Draw/src/main/java/com/yehah/draw/domain/child_work/service/ChildWorkServiceImpl.java

+++ backend/Draw/src/main/java/com/yehah/draw/domain/child_work/entity/ChildWork.java
@@ -22,7 +22,7 @@ public class ChildWork {
 
     private Long childId;
 
-    private int categoryId;
+    private Long categoryId;
 
     @Column(nullable = false)
     private Long workId; // 작업ID(동물, 내가 그린 동물, 동화 ID)
@@ -31,8 +31,9 @@ public class ChildWork {
     private String urlWork; // 작업물URL
 
     @Column(name = "url_gif", length = 200) //todo: false로 바꿔야햄
-    private String urlGif; // 작업물URL
+    private String urlGif; // 작업물GIF
 
     @Column(name = "created_date", nullable = false)
+
     private LocalDateTime createdDate; // 생성한 날짜
 }
